Fixing bugs
Getting rid of shortcut to reload init.vim; it's sort of handy, I guess;
but I found myself not using it very much, and it occasionally caused
issues when first opening a blank Neovim window. Also I've changed the
key-binding in Vifm to write changes to a file, before opening Vim. This
made it impossible to open Vifm if no buffer was currently open. So
that's gone now.