Adding to-do list
I made a markdown file to use as a to-do list a while back; and then
created a few aliases to quickly view and show the Todo list. I have no
clue why I didn't think to add the To-do list to the prompt at the time,
but it's there now.