Updating images
I've made changes to basically every aspect of the core of my dot files,
so it made since to update the screen shots on the GitHub page. The main
things that changed were my Neovim configuration; where I've built out a
completely new status bar; and my default shell, which I've cleaned-up,
and added a to-do list to the default prompt.