Adding leader key, setting up folds
I finally set up a leader key for my Vim configuration, and re-mapped a
few of my most used key-bindings to be a bit more efficient with that
power now. I ended up going with SPACE as a leader key, which means I
can launch a file explorer with SPACE+SPACE, or GOYO with SPACE+ENTER,
or navigate through splits with SPACE+VIM KEYS, so I quite happy with
that. I also set up some default setting for folds in Vim, a setting
that I don't use all that often, but like the option to have; and
finally, I adjusted some spacing in the Goyo enter and leave functions.
That change was purely cosmetic, and shouldn't effect the way anything
functions.