Switching to bash
I've been messing around with zsh and fish, but I think I'm ready to
just stick with bash for a while now. I've removed my .zshrc & .zprofile
from my dot files and replaced them with a .bashrc & a generic .profile,
that should be a bit better if I decide to switch shells again some-time
soon. I've also update the images for the repo, and the readme, to
reflect the switch to bash.