Making Alfred launch on active monitor
By default Alfred will only launch on the primary display. I thought I
really hated this feature, but it took me several months to notice I
didn't have it turned off, so maybe it's not that big of a deal. It's
fixed now though, Alacritty will launch on the active display.