Adding a key-binding to launch new window
By default opening new windows will in Alacritty is pretty weird. You
can't actually launch a new window or tab or anything else by default. I
don't mind the tabs or panes, because Tmux is a better solution anyway;
but I really need a way to launch a new widow. So I added a keybinding
to just launch a completely new Alacritty instance. It's not perfect,
but it works fine for now.