Fixing transmission-remote alias, adding a tangentely relatated script
I used to have 'tsm' aliased to run 'transmission-remote -l', which will
just list all active torrents. It turns out however that basically every
command for adding, removing, and otherwise controlling torrents uses
'transmission-remote' as a prefix. So I've dropped the '-l' on the
alias. That helped a lot. I also created a quick script to refresh the
progress of my downloaded torrents. It's a really simple for loop, but
it helps out a lot.