Adding xclip to be installed, making ddate snippet work better
I've found a great new tool called 'xclip', I like it so much that I
want it installed on every system, so now it's in my arch install
script. What I'm using it for is to add snippets to my clipboard
automatically, so that's how my ddate snippet works now. The tool is
great, it let's you put the output of any terminal command into the
clipboard.